+++ tests/e2e/data/config_with_script_responses/print_request_details.sh
@@ -1,4 +1,5 @@
 echo Server Host: $MOCK_HOST
+echo Request Host: $MOCK_REQUEST_HOST
 echo URL: $MOCK_REQUEST_URL
 echo Endpoint: $MOCK_REQUEST_ENDPOINT
 echo Method: $MOCK_REQUEST_METHOD